+++ src/main/java/com/edu/ulab/app/web/UserController.java
@@ -60,7 +60,7 @@ public class UserController {
                     @ApiResponse(description = "User book",
                             content = @Content(mediaType = MediaType.APPLICATION_JSON_VALUE,
                                     schema = @Schema(implementation = UserBookResponse.class)))})
-    public UserBookResponse getUserWithBooks(@PathVariable Long userId,
+    public UserBookResponse getUserWithBooks(@PathVariable Integer userId,
                                                 @RequestHeader(RQID) @Pattern(regexp = REQUEST_ID_PATTERN) final String requestId) {
         UserBookResponse response = userDataFacade.getUserWithBooks(userId);
         log.info("Response with user and his books: {}", response);
@@ -69,7 +69,7 @@ public class UserController {
 
     @DeleteMapping(value = "/delete/{userId}")
     @Operation(summary = "Delete user book row.")
-    public void deleteUserWithBooks(@PathVariable Long userId,
+    public void deleteUserWithBooks(@PathVariable Integer userId,
                                     @RequestHeader(RQID) @Pattern(regexp = REQUEST_ID_PATTERN) final String requestId) {
         log.info("Delete user and his books:  userId {}", userId);
         userDataFacade.deleteUserWithBooks(userId);
